Addeess code review comments
