+++ source/common/http/BUILD
@@ -301,7 +301,7 @@ envoy_cc_library(
         "//source/common/common:empty_string",
         "//source/common/common:enum_to_int",
         "//source/common/common:utility_lib",
-        "//source/common/grpc:status_lib",
+        "//source/common/grpc:utility_lib",
         "//source/common/json:json_loader_lib",
         "//source/common/network:utility_lib",
         "//source/common/protobuf:utility_lib",
