+++ source/extensions/filters/http/grpc_http1_reverse_bridge/filter.cc
@@ -3,7 +3,6 @@
 #include "common/common/enum_to_int.h"
 #include "common/grpc/codec.h"
 #include "common/grpc/common.h"
-#include "common/grpc/status.h"
 #include "common/http/headers.h"
 #include "common/http/utility.h"
 
@@ -60,7 +59,7 @@ Http::FilterHeadersStatus Filter::decodeHeaders(Http::HeaderMap& headers, bool e
   // If this is a gRPC request we:
   //  - mark this request as being gRPC
   //  - change the content-type to application/x-protobuf
-  if (Envoy::Grpc::Common::hasGrpcContentType(headers)) {
+  if (Envoy::Grpc::Utility::hasGrpcContentType(headers)) {
     enabled_ = true;
 
     // We keep track of the original content-type to ensure that we handle
